Struct aws_sdk_transcribe::client::fluent_builders::StartTranscriptionJob

pub struct StartTranscriptionJob<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> StartTranscriptionJob<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<StartTranscriptionJobOutput, SdkError<StartTranscriptionJobError>> where
R::Policy: SmithyRetryPolicy<StartTranscriptionJobInputOperationOutputAlias, StartTranscriptionJobOutput, StartTranscriptionJobError, StartTranscriptionJobInputOperationRetryAlias>,

pub fn transcription_job_name(self, inp: impl Into<String>) -> Self

The name of the job. You can't use the strings "." or ".." by themselves as the job name. The name must also be unique within an Amazon Web Services account. If you try to create a transcription job with the same name as a previous transcription job, you get a ConflictException error.

pub fn set_transcription_job_name(self, input: Option<String>) -> Self

pub fn language_code(self, inp: LanguageCode) -> Self

The language code for the language used in the input media file.

To transcribe speech in Modern Standard Arabic (ar-SA), your audio or video file must be encoded at a sample rate of 16,000 Hz or higher.

pub fn set_language_code(self, input: Option<LanguageCode>) -> Self

pub fn media_sample_rate_hertz(self, inp: i32) -> Self

The sample rate, in Hertz, of the audio track in the input media file.

If you do not specify the media sample rate, Amazon Transcribe determines the sample rate. If you specify the sample rate, it must match the sample rate detected by Amazon Transcribe. In most cases, you should leave the MediaSampleRateHertz field blank and let Amazon Transcribe determine the sample rate.

pub fn set_media_sample_rate_hertz(self, input: Option<i32>) -> Self

pub fn media_format(self, inp: MediaFormat) -> Self

The format of the input media file.

pub fn set_media_format(self, input: Option<MediaFormat>) -> Self

pub fn media(self, inp: Media) -> Self

An object that describes the input media for a transcription job.

pub fn set_media(self, input: Option<Media>) -> Self

pub fn output_bucket_name(self, inp: impl Into<String>) -> Self

The location where the transcription is stored.

If you set the OutputBucketName, Amazon Transcribe puts the transcript in the specified S3 bucket. When you call the GetTranscriptionJob operation, the operation returns this location in the TranscriptFileUri field. If you enable content redaction, the redacted transcript appears in RedactedTranscriptFileUri. If you enable content redaction and choose to output an unredacted transcript, that transcript's location still appears in the TranscriptFileUri. The S3 bucket must have permissions that allow Amazon Transcribe to put files in the bucket. For more information, see Permissions Required for IAM User Roles.

You can specify an Amazon Web Services Key Management Service (KMS) key to encrypt the output of your transcription using the OutputEncryptionKMSKeyId parameter. If you don't specify a KMS key, Amazon Transcribe uses the default Amazon S3 key for server-side encryption of transcripts that are placed in your S3 bucket.

If you don't set the OutputBucketName, Amazon Transcribe generates a pre-signed URL, a shareable URL that provides secure access to your transcription, and returns it in the TranscriptFileUri field. Use this URL to download the transcription.

pub fn set_output_bucket_name(self, input: Option<String>) -> Self

pub fn output_key(self, inp: impl Into<String>) -> Self

You can specify a location in an Amazon S3 bucket to store the output of your transcription job.

If you don't specify an output key, Amazon Transcribe stores the output of your transcription job in the Amazon S3 bucket you specified. By default, the object key is "your-transcription-job-name.json".

You can use output keys to specify the Amazon S3 prefix and file name of the transcription output. For example, specifying the Amazon S3 prefix, "folder1/folder2/", as an output key would lead to the output being stored as "folder1/folder2/your-transcription-job-name.json". If you specify "my-other-job-name.json" as the output key, the object key is changed to "my-other-job-name.json". You can use an output key to change both the prefix and the file name, for example "folder/my-other-job-name.json".

If you specify an output key, you must also specify an S3 bucket in the OutputBucketName parameter.

pub fn set_output_key(self, input: Option<String>) -> Self

pub fn output_encryption_kms_key_id(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon Web Services Key Management Service (KMS) key used to encrypt the output of the transcription job. The user calling the StartTranscriptionJob operation must have permission to use the specified KMS key.

You can use either of the following to identify a KMS key in the current account:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

KMS Key Alias: "alias/ExampleAlias"

You can use either of the following to identify a KMS key in the current account or another account:

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:region:account ID:key/1234abcd-12ab-34cd-56ef-1234567890ab"

ARN of a KMS Key Alias: "arn:aws:kms:region:account ID:alias/ExampleAlias"

If you don't specify an encryption key, the output of the transcription job is encrypted with the default Amazon S3 key (SSE-S3).

If you specify a KMS key to encrypt your output, you must also specify an output location in the OutputBucketName parameter.

pub fn set_output_encryption_kms_key_id(self, input: Option<String>) -> Self

pub fn kms_encryption_context(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

Adds a key-value pair to KMSEncryptionContext.

To override the contents of this collection use set_kms_encryption_context.

A map of plain text, non-secret key:value pairs, known as encryption context pairs, that provide an added layer of security for your data.

pub fn set_kms_encryption_context(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn settings(self, inp: Settings) -> Self

A Settings object that provides optional settings for a transcription job.

pub fn set_settings(self, input: Option<Settings>) -> Self

pub fn model_settings(self, inp: ModelSettings) -> Self

Choose the custom language model you use for your transcription job in this parameter.

pub fn set_model_settings(self, input: Option<ModelSettings>) -> Self

pub fn job_execution_settings(self, inp: JobExecutionSettings) -> Self

Provides information about how a transcription job is executed. Use this field to indicate that the job can be queued for deferred execution if the concurrency limit is reached and there are no slots available to immediately run the job.

pub fn set_job_execution_settings(
self,
input: Option<JobExecutionSettings>
) -> Self

pub fn content_redaction(self, inp: ContentRedaction) -> Self

An object that contains the request parameters for content redaction.

pub fn set_content_redaction(self, input: Option<ContentRedaction>) -> Self

pub fn identify_language(self, inp: bool) -> Self

Set this field to true to enable automatic language identification. Automatic language identification is disabled by default. You receive a BadRequestException error if you enter a value for a LanguageCode.

pub fn set_identify_language(self, input: Option<bool>) -> Self

pub fn language_options(self, inp: impl Into<LanguageCode>) -> Self

Appends an item to LanguageOptions.

To override the contents of this collection use set_language_options.

An object containing a list of languages that might be present in your collection of audio files. Automatic language identification chooses a language that best matches the source audio from that list.

To transcribe speech in Modern Standard Arabic (ar-SA), your audio or video file must be encoded at a sample rate of 16,000 Hz or higher.

pub fn set_language_options(self, input: Option<Vec<LanguageCode>>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

Add tags to an Amazon Transcribe transcription job.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self