Module aws_sdk_transcribe::error

Modules

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

internal_failure_exception

See InternalFailureException

limit_exceeded_exception

See LimitExceededException

not_found_exception

See NotFoundException

Structs

BadRequestException

Your request didn't pass one or more validation tests. For example, if the entity that you're trying to delete

ConflictException

There is already a resource with that name.

CreateCallAnalyticsCategoryError

CreateLanguageModelError

CreateMedicalVocabularyError

CreateVocabularyError

CreateVocabularyFilterError

DeleteCallAnalyticsCategoryError

DeleteCallAnalyticsJobError

DeleteLanguageModelError

DeleteMedicalTranscriptionJobError

DeleteMedicalVocabularyError

DeleteTranscriptionJobError

DeleteVocabularyError

DeleteVocabularyFilterError

DescribeLanguageModelError

GetCallAnalyticsCategoryError

GetCallAnalyticsJobError

GetMedicalTranscriptionJobError

GetMedicalVocabularyError

GetTranscriptionJobError

GetVocabularyError

GetVocabularyFilterError

InternalFailureException

There was an internal error. Check the error message and try your request again.

LimitExceededException

Either you have sent too many requests or your input file is too long. Wait before you resend your request, or

ListCallAnalyticsCategoriesError

ListCallAnalyticsJobsError

ListLanguageModelsError

ListMedicalTranscriptionJobsError

ListMedicalVocabulariesError

ListTagsForResourceError

ListTranscriptionJobsError

ListVocabulariesError

ListVocabularyFiltersError

NotFoundException

We can't find the requested resource. Check the name and try your request again.

StartCallAnalyticsJobError

StartMedicalTranscriptionJobError

StartTranscriptionJobError

TagResourceError

UntagResourceError

UpdateCallAnalyticsCategoryError

UpdateMedicalVocabularyError

UpdateVocabularyError

UpdateVocabularyFilterError

Enums

CreateCallAnalyticsCategoryErrorKind

CreateLanguageModelErrorKind

CreateMedicalVocabularyErrorKind

CreateVocabularyErrorKind

CreateVocabularyFilterErrorKind

DeleteCallAnalyticsCategoryErrorKind

DeleteCallAnalyticsJobErrorKind

DeleteLanguageModelErrorKind

DeleteMedicalTranscriptionJobErrorKind

DeleteMedicalVocabularyErrorKind

DeleteTranscriptionJobErrorKind

DeleteVocabularyErrorKind

DeleteVocabularyFilterErrorKind

DescribeLanguageModelErrorKind

GetCallAnalyticsCategoryErrorKind

GetCallAnalyticsJobErrorKind

GetMedicalTranscriptionJobErrorKind

GetMedicalVocabularyErrorKind

GetTranscriptionJobErrorKind

GetVocabularyErrorKind

GetVocabularyFilterErrorKind

ListCallAnalyticsCategoriesErrorKind

ListCallAnalyticsJobsErrorKind

ListLanguageModelsErrorKind

ListMedicalTranscriptionJobsErrorKind

ListMedicalVocabulariesErrorKind

ListTagsForResourceErrorKind

ListTranscriptionJobsErrorKind

ListVocabulariesErrorKind

ListVocabularyFiltersErrorKind

StartCallAnalyticsJobErrorKind

StartMedicalTranscriptionJobErrorKind

StartTranscriptionJobErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateCallAnalyticsCategoryErrorKind

UpdateMedicalVocabularyErrorKind

UpdateVocabularyErrorKind

UpdateVocabularyFilterErrorKind