Struct aws_sdk_transcribe::input::start_medical_transcription_job_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for StartMedicalTranscriptionJobInput

Implementations

impl Builder

pub fn medical_transcription_job_name(self, input: impl Into<String>) -> Self

The name of the medical transcription job. You can't use the strings "." or ".." by themselves as the job name. The name must also be unique within an Amazon Web Services account. If you try to create a medical transcription job with the same name as a previous medical transcription job, you get a ConflictException error.

pub fn set_medical_transcription_job_name(self, input: Option<String>) -> Self

pub fn language_code(self, input: LanguageCode) -> Self

The language code for the language spoken in the input media file. US English (en-US) is the valid value for medical transcription jobs. Any other value you enter for language code results in a BadRequestException error.

pub fn set_language_code(self, input: Option<LanguageCode>) -> Self

pub fn media_sample_rate_hertz(self, input: i32) -> Self

The sample rate, in Hertz, of the audio track in the input media file.

If you do not specify the media sample rate, Amazon Transcribe Medical determines the sample rate. If you specify the sample rate, it must match the rate detected by Amazon Transcribe Medical. In most cases, you should leave the MediaSampleRateHertz field blank and let Amazon Transcribe Medical determine the sample rate.

pub fn set_media_sample_rate_hertz(self, input: Option<i32>) -> Self

pub fn media_format(self, input: MediaFormat) -> Self

The audio format of the input media file.

pub fn set_media_format(self, input: Option<MediaFormat>) -> Self

pub fn media(self, input: Media) -> Self

Describes the input media file in a transcription request.

pub fn set_media(self, input: Option<Media>) -> Self

pub fn output_bucket_name(self, input: impl Into<String>) -> Self

The Amazon S3 location where the transcription is stored.

You must set OutputBucketName for Amazon Transcribe Medical to store the transcription results. Your transcript appears in the S3 location you specify. When you call the GetMedicalTranscriptionJob, the operation returns this location in the TranscriptFileUri field. The S3 bucket must have permissions that allow Amazon Transcribe Medical to put files in the bucket. For more information, see Permissions Required for IAM User Roles.

You can specify an Amazon Web Services Key Management Service (KMS) key to encrypt the output of your transcription using the OutputEncryptionKMSKeyId parameter. If you don't specify a KMS key, Amazon Transcribe Medical uses the default Amazon S3 key for server-side encryption of transcripts that are placed in your S3 bucket.

pub fn set_output_bucket_name(self, input: Option<String>) -> Self

pub fn output_key(self, input: impl Into<String>) -> Self

You can specify a location in an Amazon S3 bucket to store the output of your medical transcription job.

If you don't specify an output key, Amazon Transcribe Medical stores the output of your transcription job in the Amazon S3 bucket you specified. By default, the object key is "your-transcription-job-name.json".

You can use output keys to specify the Amazon S3 prefix and file name of the transcription output. For example, specifying the Amazon S3 prefix, "folder1/folder2/", as an output key would lead to the output being stored as "folder1/folder2/your-transcription-job-name.json". If you specify "my-other-job-name.json" as the output key, the object key is changed to "my-other-job-name.json". You can use an output key to change both the prefix and the file name, for example "folder/my-other-job-name.json".

If you specify an output key, you must also specify an S3 bucket in the OutputBucketName parameter.

pub fn set_output_key(self, input: Option<String>) -> Self

pub fn output_encryption_kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon Web Services Key Management Service (KMS) key used to encrypt the output of the transcription job. The user calling the StartMedicalTranscriptionJob operation must have permission to use the specified KMS key.

You use either of the following to identify a KMS key in the current account:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

KMS Key Alias: "alias/ExampleAlias"

You can use either of the following to identify a KMS key in the current account or another account:

Amazon Resource Name (ARN) of a KMS key in the current account or another account: "arn:aws:kms:region:account ID:key/1234abcd-12ab-34cd-56ef-1234567890ab"

ARN of a KMS Key Alias: "arn:aws:kms:region:account ID:alias/ExampleAlias"

If you don't specify an encryption key, the output of the medical transcription job is encrypted with the default Amazon S3 key (SSE-S3).

If you specify a KMS key to encrypt your output, you must also specify an output location in the OutputBucketName parameter.

pub fn set_output_encryption_kms_key_id(self, input: Option<String>) -> Self

pub fn kms_encryption_context(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

pub fn set_kms_encryption_context(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn settings(self, input: MedicalTranscriptionSetting) -> Self

Optional settings for the medical transcription job.

pub fn set_settings(self, input: Option<MedicalTranscriptionSetting>) -> Self

pub fn content_identification_type(
self,
input: MedicalContentIdentificationType
) -> Self

You can configure Amazon Transcribe Medical to label content in the transcription output. If you specify PHI, Amazon Transcribe Medical labels the personal health information (PHI) that it identifies in the transcription output.

pub fn set_content_identification_type(
self,
input: Option<MedicalContentIdentificationType>
) -> Self

pub fn specialty(self, input: Specialty) -> Self

The medical specialty of any clinician speaking in the input media.

pub fn set_specialty(self, input: Option<Specialty>) -> Self

pub fn type(self, input: Type) -> Self

The type of speech in the input audio. CONVERSATION refers to conversations between two or more speakers, e.g., a conversations between doctors and patients. DICTATION refers to single-speaker dictated speech, such as clinical notes.

pub fn set_type(self, input: Option<Type>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn build(self) -> Result<StartMedicalTranscriptionJobInput, BuildError>

Consumes the builder and constructs a StartMedicalTranscriptionJobInput