Struct aws_sdk_transcribe::input::StartCallAnalyticsJobInput

#[non_exhaustive]
pub struct StartCallAnalyticsJobInput {
    pub call_analytics_job_name: Option<String>,
    pub media: Option<Media>,
    pub output_location: Option<String>,
    pub output_encryption_kms_key_id: Option<String>,
    pub data_access_role_arn: Option<String>,
    pub settings: Option<CallAnalyticsJobSettings>,
    pub channel_definitions: Option<Vec<ChannelDefinition>>,
}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

call_analytics_job_name: Option<String>

The name of the call analytics job. You can't use the string "." or ".." by themselves as the job name. The name must also be unique within an Amazon Web Services account. If you try to create a call analytics job with the same name as a previous call analytics job, you get a ConflictException error.

media: Option<Media>

Describes the input media file in a transcription request.

output_location: Option<String>

The Amazon S3 location where the output of the call analytics job is stored. You can provide the following location types to store the output of call analytics job:

s3://DOC-EXAMPLE-BUCKET1

If you specify a bucket, Amazon Transcribe saves the output of the analytics job as a JSON file at the root level of the bucket.

s3://DOC-EXAMPLE-BUCKET1/folder/

f you specify a path, Amazon Transcribe saves the output of the analytics job as s3://DOC-EXAMPLE-BUCKET1/folder/your-transcription-job-name.json

If you specify a folder, you must provide a trailing slash.

s3://DOC-EXAMPLE-BUCKET1/folder/filename.json

If you provide a path that has the filename specified, Amazon Transcribe saves the output of the analytics job as s3://DOC-EXAMPLEBUCKET1/folder/filename.json

You can specify an Amazon Web Services Key Management Service (KMS) key to encrypt the output of our analytics job using the OutputEncryptionKMSKeyId parameter. If you don't specify a KMS key, Amazon Transcribe uses the default Amazon S3 key for server-side encryption of the analytics job output that is placed in your S3 bucket.

output_encryption_kms_key_id: Option<String>

The Amazon Resource Name (ARN) of the Amazon Web Services Key Management Service key used to encrypt the output of the call analytics job. The user calling the operation must have permission to use the specified KMS key.

You use either of the following to identify an Amazon Web Services KMS key in the current account:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

KMS Key Alias: "alias/ExampleAlias"

You can use either of the following to identify a KMS key in the current account or another account:

Amazon Resource Name (ARN) of a KMS key in the current account or another account: "arn:aws:kms:region:account ID:key/1234abcd-12ab-34cd-56ef1234567890ab"

ARN of a KMS Key Alias: "arn:aws:kms:region:account ID:alias/ExampleAlias"

If you don't specify an encryption key, the output of the call analytics job is encrypted with the default Amazon S3 key (SSE-S3).

If you specify a KMS key to encrypt your output, you must also specify an output location in the OutputLocation parameter.

data_access_role_arn: Option<String>

The Amazon Resource Name (ARN) of a role that has access to the S3 bucket that contains your input files. Amazon Transcribe assumes this role to read queued audio files. If you have specified an output S3 bucket for your transcription results, this role should have access to the output bucket as well.

settings: Option<CallAnalyticsJobSettings>

A Settings object that provides optional settings for a call analytics job.

channel_definitions: Option<Vec<ChannelDefinition>>

When you start a call analytics job, you must pass an array that maps the agent and the customer to specific audio channels. The values you can assign to a channel are 0 and 1. The agent and the customer must each have their own channel. You can't assign more than one channel to an agent or customer.

Implementations

impl StartCallAnalyticsJobInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<StartCallAnalyticsJob, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<StartCallAnalyticsJob>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture StartCallAnalyticsJobInput