Module aws_sdk_transcribe::model

Modules

absolute_time_range

See AbsoluteTimeRange

call_analytics_job

See CallAnalyticsJob

call_analytics_job_settings

See CallAnalyticsJobSettings

call_analytics_job_summary

See CallAnalyticsJobSummary

category_properties

See CategoryProperties

channel_definition

See ChannelDefinition

content_redaction

See ContentRedaction

input_data_config

See InputDataConfig

interruption_filter

See InterruptionFilter

job_execution_settings

See JobExecutionSettings

language_model

See LanguageModel

media

See Media

medical_transcript

See MedicalTranscript

medical_transcription_job

See MedicalTranscriptionJob

medical_transcription_job_summary

See MedicalTranscriptionJobSummary

medical_transcription_setting

See MedicalTranscriptionSetting

model_settings

See ModelSettings

non_talk_time_filter

See NonTalkTimeFilter

relative_time_range

See RelativeTimeRange

sentiment_filter

See SentimentFilter

settings

See Settings

tag

See Tag

transcript

See Transcript

transcript_filter

See TranscriptFilter

transcription_job

See TranscriptionJob

transcription_job_summary

See TranscriptionJobSummary

vocabulary_filter_info

See VocabularyFilterInfo

vocabulary_info

See VocabularyInfo

Structs

AbsoluteTimeRange

A time range, set in seconds, between two points in the call.

CallAnalyticsJob

Describes an asynchronous analytics job that was created with the StartAnalyticsJob

CallAnalyticsJobSettings

Provides optional settings for the CallAnalyticsJob operation.

CallAnalyticsJobSummary

Provides summary information about a call analytics job.

CategoryProperties

An object that contains the rules and additional information about a call analytics category.

ChannelDefinition

For a call analytics job, an object that indicates the audio channel that belongs to the agent and the audio

ContentRedaction

Settings for content redaction within a transcription job.

InputDataConfig

The object that contains the Amazon S3 object location and access role required to train and tune your custom

InterruptionFilter

An object that enables you to configure your category to be applied to call analytics jobs where either the

JobExecutionSettings

Provides information about when a transcription job should be executed.

LanguageModel

The structure used to describe a custom language model.

Media

Describes the input media file in a transcription request.

MedicalTranscript

Identifies the location of a medical transcript.

MedicalTranscriptionJob

The data structure that contains the information for a medical transcription job.

MedicalTranscriptionJobSummary

Provides summary information about a transcription job.

MedicalTranscriptionSetting

Optional settings for the StartMedicalTranscriptionJob

ModelSettings

The object used to call your custom language model to your transcription job.

NonTalkTimeFilter

An object that enables you to configure your category to be applied to call analytics jobs where either the

RelativeTimeRange

An object that allows percentages to specify the proportion of the call where you would like to apply a filter. For

SentimentFilter

An object that enables you to specify a particular customer or agent sentiment. If at least 50 percent of the

Settings

Provides optional settings for the StartTranscriptionJob operation.

Tag

A key:value pair that adds metadata to a resource used by Amazon Transcribe. For example, a tag with the key:value pair

Transcript

Identifies the location of a transcription.

TranscriptFilter

Matches the output of the transcription to either the specific phrases that you specify, or the intent of the phrases

TranscriptionJob

Describes an asynchronous transcription job that was created with the StartTranscriptionJob

TranscriptionJobSummary

Provides a summary of information about a transcription job.

VocabularyFilterInfo

Provides information about a vocabulary filter.

VocabularyInfo

Provides information about a custom vocabulary.

Enums

BaseModelName

CallAnalyticsJobStatus

ClmLanguageCode

LanguageCode

MediaFormat

MedicalContentIdentificationType

ModelStatus

OutputLocationType

ParticipantRole

RedactionOutput

RedactionType

Rule

SentimentValue

Specialty

TranscriptFilterType

TranscriptionJobStatus

Type

VocabularyFilterMethod

VocabularyState