Struct aws_sdk_transcribe::model::medical_transcription_job::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for MedicalTranscriptionJob

Implementations

impl Builder

pub fn medical_transcription_job_name(self, input: impl Into<String>) -> Self

The name for a given medical transcription job.

pub fn set_medical_transcription_job_name(self, input: Option<String>) -> Self

pub fn transcription_job_status(self, input: TranscriptionJobStatus) -> Self

The completion status of a medical transcription job.

pub fn set_transcription_job_status(
self,
input: Option<TranscriptionJobStatus>
) -> Self

pub fn language_code(self, input: LanguageCode) -> Self

The language code for the language spoken in the source audio file. US English (en-US) is the only supported language for medical transcriptions. Any other value you enter for language code results in a BadRequestException error.

pub fn set_language_code(self, input: Option<LanguageCode>) -> Self

pub fn media_sample_rate_hertz(self, input: i32) -> Self

The sample rate, in Hertz, of the source audio containing medical information.

If you don't specify the sample rate, Amazon Transcribe Medical determines it for you. If you choose to specify the sample rate, it must match the rate detected by Amazon Transcribe Medical. In most cases, you should leave the MedicalMediaSampleHertz blank and let Amazon Transcribe Medical determine the sample rate.

pub fn set_media_sample_rate_hertz(self, input: Option<i32>) -> Self

pub fn media_format(self, input: MediaFormat) -> Self

The format of the input media file.

pub fn set_media_format(self, input: Option<MediaFormat>) -> Self

pub fn media(self, input: Media) -> Self

Describes the input media file in a transcription request.

pub fn set_media(self, input: Option<Media>) -> Self

pub fn transcript(self, input: MedicalTranscript) -> Self

An object that contains the MedicalTranscript. The MedicalTranscript contains the TranscriptFileUri.

pub fn set_transcript(self, input: Option<MedicalTranscript>) -> Self

pub fn start_time(self, input: Instant) -> Self

A timestamp that shows when the job started processing.

pub fn set_start_time(self, input: Option<Instant>) -> Self

pub fn creation_time(self, input: Instant) -> Self

A timestamp that shows when the job was created.

pub fn set_creation_time(self, input: Option<Instant>) -> Self

pub fn completion_time(self, input: Instant) -> Self

A timestamp that shows when the job was completed.

pub fn set_completion_time(self, input: Option<Instant>) -> Self

pub fn failure_reason(self, input: impl Into<String>) -> Self

If the TranscriptionJobStatus field is FAILED, this field contains information about why the job failed.

The FailureReason field contains one of the following values:

Unsupported media format- The media format specified in the MediaFormat field of the request isn't valid. See the description of the MediaFormat field for a list of valid values.

The media format provided does not match the detected media format- The media format of the audio file doesn't match the format specified in the MediaFormat field in the request. Check the media format of your media file and make sure the two values match.

Invalid sample rate for audio file- The sample rate specified in the MediaSampleRateHertz of the request isn't valid. The sample rate must be between 8,000 and 48,000 Hertz.

The sample rate provided does not match the detected sample rate- The sample rate in the audio file doesn't match the sample rate specified in the MediaSampleRateHertz field in the request. Check the sample rate of your media file and make sure that the two values match.

Invalid file size: file size too large- The size of your audio file is larger than what Amazon Transcribe Medical can process. For more information, see Guidelines and Quotas in the Amazon Transcribe Medical Guide

Invalid number of channels: number of channels too large- Your audio contains more channels than Amazon Transcribe Medical is configured to process. To request additional channels, see Amazon Transcribe Medical Endpoints and Quotas in the Amazon Web Services General Reference

pub fn set_failure_reason(self, input: Option<String>) -> Self

pub fn settings(self, input: MedicalTranscriptionSetting) -> Self

Object that contains object.

pub fn set_settings(self, input: Option<MedicalTranscriptionSetting>) -> Self

pub fn content_identification_type(
self,
input: MedicalContentIdentificationType
) -> Self

Shows the type of content that you've configured Amazon Transcribe Medical to identify in a transcription job. If the value is PHI, you've configured the job to identify personal health information (PHI) in the transcription output.

pub fn set_content_identification_type(
self,
input: Option<MedicalContentIdentificationType>
) -> Self

pub fn specialty(self, input: Specialty) -> Self

The medical specialty of any clinicians providing a dictation or having a conversation. Refer to Transcribing a medical conversationfor a list of supported specialties.

pub fn set_specialty(self, input: Option<Specialty>) -> Self

pub fn type(self, input: Type) -> Self

The type of speech in the transcription job. CONVERSATION is generally used for patient-physician dialogues. DICTATION is the setting for physicians speaking their notes after seeing a patient. For more information, see What is Amazon Transcribe Medical?.

pub fn set_type(self, input: Option<Type>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn build(self) -> MedicalTranscriptionJob

Consumes the builder and constructs a MedicalTranscriptionJob