Struct aws_sdk_transcribe::model::CallAnalyticsJob

#[non_exhaustive]
pub struct CallAnalyticsJob {

    pub call_analytics_job_name: Option<String>,
    pub call_analytics_job_status: Option<CallAnalyticsJobStatus>,
    pub language_code: Option<LanguageCode>,
    pub media_sample_rate_hertz: Option<i32>,
    pub media_format: Option<MediaFormat>,
    pub media: Option<Media>,
    pub transcript: Option<Transcript>,
    pub start_time: Option<Instant>,
    pub creation_time: Option<Instant>,
    pub completion_time: Option<Instant>,
    pub failure_reason: Option<String>,
    pub data_access_role_arn: Option<String>,
    pub identified_language_score: Option<f32>,
    pub settings: Option<CallAnalyticsJobSettings>,
    pub channel_definitions: Option<Vec<ChannelDefinition>>,

}

Describes an asynchronous analytics job that was created with the StartAnalyticsJob operation.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

call_analytics_job_name: Option<String>

The name of the call analytics job.

call_analytics_job_status: Option<CallAnalyticsJobStatus>

The status of the analytics job.

language_code: Option<LanguageCode>

If you know the language spoken between the customer and the agent, specify a language code for this field.

If you don't know the language, you can leave this field blank, and Amazon Transcribe will use machine learning to automatically identify the language. To improve the accuracy of language identification, you can provide an array containing the possible language codes for the language spoken in your audio. Refer to Supported languages and language-specific features for additional information.

media_sample_rate_hertz: Option<i32>

The sample rate, in Hertz, of the audio.

media_format: Option<MediaFormat>

The format of the input audio file. Note: for call analytics jobs, only the following media formats are supported: MP3, MP4, WAV, FLAC, OGG, and WebM.

media: Option<Media>

Describes the input media file in a transcription request.

transcript: Option<Transcript>

Identifies the location of a transcription.

start_time: Option<Instant>

A timestamp that shows when the analytics job started processing.

creation_time: Option<Instant>

A timestamp that shows when the analytics job was created.

completion_time: Option<Instant>

A timestamp that shows when the analytics job was completed.

failure_reason: Option<String>

If the AnalyticsJobStatus is FAILED, this field contains information about why the job failed.

The FailureReason field can contain one of the following values:

Unsupported media format: The media format specified in the MediaFormat field of the request isn't valid. See the description of the MediaFormat field for a list of valid values.

The media format provided does not match the detected media format: The media format of the audio file doesn't match the format specified in the MediaFormat field in the request. Check the media format of your media file and make sure the two values match.

Invalid sample rate for audio file: The sample rate specified in the MediaSampleRateHertz of the request isn't valid. The sample rate must be between 8,000 and 48,000 Hertz.

The sample rate provided does not match the detected sample rate: The sample rate in the audio file doesn't match the sample rate specified in the MediaSampleRateHertz field in the request. Check the sample rate of your media file and make sure that the two values match.

Invalid file size: file size too large: The size of your audio file is larger than what Amazon Transcribe Medical can process. For more information, see Guidelines and Quotas in the Amazon Transcribe Medical Guide.

Invalid number of channels: number of channels too large: Your audio contains more channels than Amazon Transcribe Medical is configured to process. To request additional channels, see Amazon Transcribe Medical Endpoints and Quotas in the Amazon Web Services General Reference.

data_access_role_arn: Option<String>

The Amazon Resource Number (ARN) that you use to get access to the analytics job.

identified_language_score: Option<f32>

A value between zero and one that Amazon Transcribe assigned to the language that it identified in the source audio. This value appears only when you don't provide a single language code. Larger values indicate that Amazon Transcribe has higher confidence in the language that it identified

settings: Option<CallAnalyticsJobSettings>

Provides information about the settings used to run a transcription job.

channel_definitions: Option<Vec<ChannelDefinition>>

Shows numeric values to indicate the channel assigned to the agent's audio and the channel assigned to the customer's audio.

Implementations

impl CallAnalyticsJob

pub fn builder() -> Builder

Creates a new builder-style object to manufacture CallAnalyticsJob