Module aws_sdk_transcribe::operation

Structs

CreateCallAnalyticsCategory

Creates an analytics category. Amazon Transcribe applies the conditions specified by your

CreateLanguageModel

Creates a new custom language model. Use Amazon S3 prefixes to provide the location of your input files. The time it

CreateMedicalVocabulary

Creates a new custom vocabulary that you can use to modify how Amazon Transcribe Medical transcribes your audio file.

CreateVocabulary

Creates a new custom vocabulary that you can use to change the way Amazon Transcribe handles transcription of an

CreateVocabularyFilter

Creates a new vocabulary filter that you can use to filter words, such as profane words, from the output of

DeleteCallAnalyticsCategory

Deletes a call analytics category using its name.

DeleteCallAnalyticsJob

Deletes a call analytics job using its name.

DeleteLanguageModel

Deletes a custom language model using its name.

DeleteMedicalTranscriptionJob

Deletes a transcription job generated by Amazon Transcribe Medical and any related information.

DeleteMedicalVocabulary

Deletes a vocabulary from Amazon Transcribe Medical.

DeleteTranscriptionJob

Deletes a previously submitted transcription job along with any other generated results such as the

DeleteVocabulary

Deletes a vocabulary from Amazon Transcribe.

DeleteVocabularyFilter

Removes a vocabulary filter.

DescribeLanguageModel

Gets information about a single custom language model. Use this information to see details about the

GetCallAnalyticsCategory

Retrieves information about a call analytics category.

GetCallAnalyticsJob

Returns information about a call analytics job. To see the status of the job, check the

GetMedicalTranscriptionJob

Returns information about a transcription job from Amazon Transcribe Medical. To see the status of the job, check the

GetMedicalVocabulary

Retrieves information about a medical vocabulary.

GetTranscriptionJob

Returns information about a transcription job. To see the status of the job, check the

GetVocabulary

Gets information about a vocabulary.

GetVocabularyFilter

Returns information about a vocabulary filter.

ListCallAnalyticsCategories

Provides more information about the call analytics categories that you've created. You

ListCallAnalyticsJobs

List call analytics jobs with a specified status or substring that matches their names.

ListLanguageModels

Provides more information about the custom language models you've created. You can use the information in

ListMedicalTranscriptionJobs

Lists medical transcription jobs with a specified status or substring that matches their names.

ListMedicalVocabularies

Returns a list of vocabularies that match the specified criteria. If you don't enter a value in any of the request

ListTagsForResource

Lists all tags associated with a given transcription job, vocabulary, or resource.

ListTranscriptionJobs

Lists transcription jobs with the specified status.

ListVocabularies

Returns a list of vocabularies that match the specified criteria. If no criteria are specified, returns the entire list

ListVocabularyFilters

Gets information about vocabulary filters.

StartCallAnalyticsJob

Starts an asynchronous analytics job that not only transcribes the audio recording of a caller and agent, but

StartMedicalTranscriptionJob

Starts a batch job to transcribe medical speech to text.

StartTranscriptionJob

Starts an asynchronous job to transcribe speech to text.

TagResource

Tags a Amazon Transcribe resource with the given list of tags.

UntagResource

Removes specified tags from a specified Amazon Transcribe resource.

UpdateCallAnalyticsCategory

Updates the call analytics category with new values. The UpdateCallAnalyticsCategory

UpdateMedicalVocabulary

Updates a vocabulary with new values that you provide in a different text file from the one you used to create

UpdateVocabulary

Updates an existing vocabulary with new values. The UpdateVocabulary operation overwrites

UpdateVocabularyFilter

Updates a vocabulary filter with a new list of filtered words.